Merge pull request #6 from Vladooha/master

+++ src/main/java/com/vladooha/data/repositories/LoginInfoRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LoginInfoRepo extends JpaRepository<LoginInfo, Integer> {
+public interface LoginInfoRepo extends JpaRepository<LoginInfo, Long> {
     LoginInfo findByUsername(String username);
     LoginInfo findByEmail(String email);
     List<LoginInfo> findByUsernameOrEmail(String username, String email);
